% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{capitulos.productos}
\alias{capitulos.productos}
\title{Capítulos de productos}
\format{Un data-frame con dos variables: \code{capitulo} y \code{descripcion_capitulo}.
(\code{capitulos.productos} está desagregada en las 99 capítulos de productos, cada uno con su respectiva
  descripción)}
\usage{
capitulos.productos
}
\description{
A partir de codiguera de NCM.
}
\examples{
capitulos.productos
}
\keyword{datasets}
