% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ciiu}
\alias{ciiu}
\title{Listado códigos CIIU del INE}
\format{Un data-frame con once variables:
\describe{
\item{\code{seccion}}{seccion de actividad}
\item{\code{ciiu_4}}{código ciiu a 4 dígitos}
\item{\code{descripcion}}{descripción de \code{ciiu_4}}
\item{\code{ciiu}}{código ciiu a 5 dígitos}
\item{\code{division}}{division de actividad}
\item{\code{grupo}}{grupo de actividad}
\item{\code{descripcion_grupo}}{descripción de \code{grupo}}
\item{\code{clase}}{clase de actividad}
\item{\code{descripcion_clase}}{descripción de \code{clase}}
\item{\code{subclase}}{subclase de actividad}
\item{\code{descripcion_subclase}}{descripción de \code{subclase}}
}

(\code{ciiu} está desagregada
  en los 925 códigos disponibles a 5 dígitos, y sus correspondientes código ciiu
  a 4 dígitos, clases, subclases, grupos y secciones, cada uno con su respectiva
  descripción).}
\usage{
ciiu
}
\description{
Análisis codiguera del INE.
}
\examples{
  ciiu
}
\keyword{datasets}
