% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ciiu4_ciiu3}
\alias{ciiu4_ciiu3}
\title{ciiu4_ciiu3}
\format{Un data-frame con dos variables: \code{ciiu3} y \code{ciiu4}.
(\code{ciiu4_ciiu3} está desagregada en 440 relaciones ciiu3 - ciiu4)}
\usage{
ciiu4_ciiu3
}
\description{
Equivalencia entre Clasificación Industrial Internacional Uniforme Rev.3 y Rev.4.
}
\examples{
ciiu4_ciiu3
}
\keyword{datasets}
