% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clase}
\alias{clase}
\title{Clases de actividad}
\format{Un data-frame con dos variables: \code{clase} y \code{desc_clase}.
(\code{clase} está desagregada en las 286 clases de actividad, cada una con su respectiva
  descripción)}
\usage{
clase
}
\description{
A partir de codiguera del INE.
}
\examples{
clase
}
\keyword{datasets}
