
\docType{data}
\name{clasif.uyxxi}
\alias{clasif.uyxxi}
\title{Codiguera Uruguay XXI}
\format{Un data-frame con cuatro variables:
\describe{
\item{\code{ncm_4}}{Código de producto a 4 dígitos}
\item{\code{ncm_6}}{Código de producto a 6 dígitos}
\item{\code{desc.uyxxi}}{Nombre del grupo de Uruguay XXI}
\item{\code{capitulo}}{Capítulo del producto}
}

(\code{clasif.uyxxi} está desagregada
  en 3289 productos a 6 dígitos, cada uno con su respectiva
  grupo de clasificación de Uruguay XXI).}
\usage{
clasif.uyxxi
}
\description{
Codiguera de grupos de productos de Uruguay XXI.
}
\examples{
  clasif.uyxxi
}
\keyword{datasets}
