% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnbcu.completo}
\alias{cnbcu.completo}
\title{cnbcu.completo}
\format{Un data-frame con cuatro variables: \code{desc.cnbcu}, \code{year}, \code{indice} y \code{cnbcu}.
(\code{cnbcu.completo} está desagregada en 1044 códigos de actividad del BCU, cada uno con su respectiva descripción)}
\usage{
cnbcu.completo
}
\description{
Codiguera completa BCU desagregada por código de actividad (cnbcu), con descripción.
}
\examples{
cnbcu.completo
}
\keyword{datasets}
