% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnbcu.seccion}
\alias{cnbcu.seccion}
\title{cnbcu.seccion}
\format{Un data-frame con cuatro variables: \code{desc.cnbcu.seccion}, \code{year}, \code{indice.seccion}
y \code{cnbcu.seccion}.
(\code{cnbcu.seccion} está desagregada en 16 códigos de secciones de actividad del BCU, cada uno con su respectiva descripción para
12 años de estudio)}
\usage{
cnbcu.seccion
}
\description{
Codiguera completa BCU desagregada por subsección, con descripción.
}
\examples{
cnbcu.seccion
}
\keyword{datasets}
