% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnbcu.subseccion}
\alias{cnbcu.subseccion}
\title{cnbcu.subseccion}
\format{Un data-frame con cuatro variables: \code{desc.cnbcu.subseccion}, \code{year}, \code{indice.subseccion}
y \code{cnbcu.subseccion}.
(\code{cnbcu.subseccion} está desagregada en 852 códigos de subsecciones de actividad del BCU, cada uno con su respectiva descripción y
  correspondencia CIIU Rev.3)}
\usage{
cnbcu.subseccion
}
\description{
Codiguera completa BCU desagregada por subsección, con descripción.
}
\examples{
cnbcu.subseccion
}
\keyword{datasets}
