% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnbcu_ciiu}
\alias{cnbcu_ciiu}
\title{cnbcu_ciiu}
\format{Un data-frame con cuatro variables: \code{CATEGORIA}, \code{`CODIGO CNBCU`}, \code{`DENOMINACION CNBCU`}
y \code{`CODIGO CIIU Rev.3`}.
(\code{cnbcu_ciiu} est치 desagregada en 177 c칩digos de actividad del BCU, cada uno con su respectiva descripci칩n y
  correspondencia CIIU Rev.3)}
\usage{
cnbcu_ciiu
}
\description{
Codiguera completa BCU con descripci칩n y equivalencia ciiu3.
}
\examples{
cnbcu_ciiu
}
\keyword{datasets}
