% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnbcu_subclase}
\alias{cnbcu_subclase}
\title{cnbcu_subclase}
\format{Un data-frame con tres variables: \code{cnbcu}, \code{denominacion.bcu} y \code{ciiu3}.
(\code{cnbcu_subclase} está desagregada en 301 códigos de cnbcu, cada uno con su respectiva
  descripción y equivalencia de Clasificación Industrial Uniforme Rev.3.)}
\usage{
cnbcu_subclase
}
\description{
Codiguera completa BCU con descripción y equivalencia ciiu3.
}
\examples{
cnbcu_subclase
}
\keyword{datasets}
