% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{correspondencia_ciiu4_ciiu3}
\alias{correspondencia_ciiu4_ciiu3}
\title{Listado correspondencia CIIU Rev.4 con CIIU Rev.3 del INE}
\format{Un data-frame con cuatro variables:
\describe{
\item{\code{ciiu3}}{Clase de actividad CIIU Rev.3}
\item{\code{ciiu4_5}}{Subclase de actividad CIIU Rev.4}
\item{\code{ciiu4}}{Clase de actividad CIIU Rev.4}
\item{\code{descripcion}}{descripción de \code{ciiu4_5}}
}
(\code{ciiu4_5} está desagregada
  787 códigos de subclase de actividad, cada uno con su respectiva descripción y
  correspondencia con \code{ciiu4} y \code{ciiu3}).}
\usage{
correspondencia_ciiu4_ciiu3
}
\description{
Análisis correspondencia CIIU Rev.4 con CIIU Rev.3 del INE.
}
\examples{
  correspondencia_ciiu4_ciiu3
}
\keyword{datasets}
