% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{division}
\alias{division}
\title{Divisiones de actividad}
\format{Un data-frame con dos variables: \code{division} y \code{desc_division}.
(\code{division} está desagregada en las 88 divisiones de actividad, cada uno con su respectiva
  descripción)}
\usage{
division
}
\description{
A partir de codiguera del INE.
}
\examples{
division
}
\keyword{datasets}
