% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grupo}
\alias{grupo}
\title{Grupos de actividad}
\format{Un data-frame con dos variables: \code{grupo} y \code{desc_grupo}.
(\code{grupo} está desagregada en los 235 grupos de actividad, cada uno con su respectiva
  descripción)}
\usage{
grupo
}
\description{
A partir de codiguera del INE.
}
\examples{
grupo
}
\keyword{datasets}
