% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{listado_ocupaciones}
\alias{listado_ocupaciones}
\title{Listado de Ocupaciones}
\format{Un data-frame con cuatro variables: \code{`Listado Nacional de Denominaciones de Ocupaciones`},
\code{`Código CNUO 95`}, \code{`Descripción de los Grupos Primarios de la CIUO 08`} y \code{`Código CIUO 08`}.
(\code{listado_ocupaciones} está desagregada en 3509 ocupaciones, cada una con su respectiva
  descripción)}
\usage{
listado_ocupaciones
}
\description{
Listado Nacional de Denominaciones de Ocupaciones (INE).
}
\examples{
listado_ocupaciones
}
\keyword{datasets}
