% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NCM}
\alias{NCM}
\title{NCM}
\format{Un data-frame con tres variables: \code{ncm_4}, \code{descripcionNCM} y \code{capitulo}.
(\code{NCM} está desagregada en 1226 códigos de productos a 4 dígitos, cada uno con su respectiva
  descripción)}
\usage{
NCM
}
\description{
Listado de productos Nomenclatura Común Mercosur
}
\examples{
NCM
}
\keyword{datasets}
