% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ncm_base}
\alias{ncm_base}
\title{ncm_base}
\format{Un data-frame con cinco variables: \code{NCM}, \code{descripcion}, \code{codigo},
\code{n} y \code{variable}.
(\code{ncm_base} está desagregada en 14542 códigos de productos a 10 dígitos, cada uno con su respectiva
  descripción)}
\usage{
ncm_base
}
\description{
Codiguera completa NCM adaptada a Uruguay
}
\examples{
ncm_base
}
\keyword{datasets}
