% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seccion}
\alias{seccion}
\title{Secciones de actividad}
\format{Un data-frame con dos variables: \code{seccion} y \code{desc_seccion}.
(\code{seccion} está desagregada en las 21 secciones de actividad, cada una con su respectiva
  descripción)}
\usage{
seccion
}
\description{
A partir de codiguera del INE.
}
\examples{
seccion
}
\keyword{datasets}
