% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seccion_division}
\alias{seccion_division}
\title{Sección - Division}
\format{Un data-frame con dos variables: \code{seccion} y \code{division}.
(\code{seccion} está desagregada en las 21 secciones de actividad disponibles, y en cada caso
las divisiones que le corresponden)}
\usage{
seccion_division
}
\description{
A partir de codiguera del INE.
}
\examples{
seccion_division
}
\keyword{datasets}
