% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{secciones.productos}
\alias{secciones.productos}
\title{Secciones de productos}
\format{Un data-frame con dos variables: \code{seccion} y \code{descripcion_seccion}.
(\code{secciones.productos} está desagregada en las 21 secciones de productos, cada una con su respectiva
  descripción)}
\usage{
secciones.productos
}
\description{
A partir de codiguera de NCM.
}
\examples{
secciones.productos
}
\keyword{datasets}
