% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{subclase}
\alias{subclase}
\title{Subclases de actividad}
\format{Un data-frame con dos variables: \code{subclase} y \code{desc_subclase}.
(\code{subclase} está desagregada en las 286 subclases de actividad, cada una con su respectiva
  descripción)}
\usage{
subclase
}
\description{
A partir de codiguera del INE.
}
\examples{
subclase
}
\keyword{datasets}
